ADD: preparation for images larger than 64k resp 255*255
git-svn-id: svn://tron.homeunix.org/simutrans/simutrans/trunk@3174 8aca7d54-2c30-db11-9de9-000461428c89

+++ besch/obj_node_info.h
@@ -19,7 +19,7 @@
 struct obj_node_info_t {
     uint32  type;
     uint16  children;
-    uint16  size;
+    uint32  size;
 };
 
 #endif
